Deprecate `SafeSignerLaunchpad`
This PR removes the `SafeSignerLaunchpad` as we can provide a better
developer experience with just the `SafeWebAuthnSharedSigner` contract
when working with ERC-4337.
