+++ examples/4337-passkeys-singleton-signer/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@safe-global/safe-modules-example-4337-passkeys-singleton-signer",
+  "name": "@safe-global/safe-modules-example-4337-passkeys",
   "private": true,
   "version": "0.0.0",
   "type": "module",
@@ -13,7 +13,8 @@
   "dependencies": {
     "@account-abstraction/contracts": "0.7.0",
     "@safe-global/safe-4337": "0.3.0",
-    "@safe-global/safe-passkey": "0.2.0-alpha.1",
+    "@safe-global/safe-contracts": "^1.4.1-build.0",
+    "@safe-global/safe-passkey": "workspace:0.2.0-alpha.2",
     "@web3modal/ethers": "^4.1.11",
     "ethers": "^6.12.1",
     "react": "^18.3.1",
