--- examples/4337-passkeys/.env.example
@@ -1,5 +1,0 @@
-// Get projectId at https://cloud.walletconnect.com
-VITE_WC_CLOUD_PROJECT_ID=
-// 4337 Bundler URL. We recommend https://www.pimlico.io/
-VITE_WC_4337_BUNDLER_URL=
-
