--- examples/4337-passkeys/.eslintrc.cjs
@@ -1,14 +1,0 @@
-module.exports = {
-  ignorePatterns: ['dist', '.eslintrc.cjs'],
-  extends: ['../../.eslintrc.js', 'plugin:react-hooks/recommended'],
-  plugins: ['react-refresh'],
-  rules: {
-    'react-refresh/only-export-components': [
-      'warn',
-      {
-        allowExportNames: ['meta', 'links', 'headers', 'loader', 'action'],
-        allowConstantExport: true,
-      },
-    ],
-  },
-};
