--- examples/4337-passkeys/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@safe-global/safe-modules-example-4337-passkeys",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "lint:fix": "pnpm run lint -- --fix",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@account-abstraction/contracts": "0.7.0",
-    "@safe-global/safe-4337": "0.3.0",
-    "@safe-global/safe-passkey": "workspace:0.2.0-alpha.2",
-    "@web3modal/ethers": "^4.1.11",
-    "ethers": "^6.12.1",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-router-dom": "^6.23.1"
-  },
-  "devDependencies": {
-    "@types/react": "^18.3.2",
-    "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react-swc": "^3.6.0",
-    "react-router": "^6.23.1",
-    "typescript": "^5.4.5",
-    "vite": "^5.2.11",
-    "vite-plugin-commonjs": "^0.10.1"
-  }
-}
