--- examples/4337-passkeys/src/components/ConnectButton.tsx
@@ -1,3 +1,0 @@
-export default function ConnectButton() {
-  return <w3m-button />
-}
