--- examples/4337-passkeys/src/components/ConnectWallet.tsx
@@ -1,9 +1,0 @@
-function ConnectWallet() {
-  return (
-    <div className="card">
-      <p>Please follow the instructions to connect your wallet by clicking the "Connect Wallet" button.</p>
-    </div>
-  )
-}
-
-export { ConnectWallet }
