--- examples/4337-passkeys/src/hooks/UseOutletContext.tsx
@@ -1,12 +1,0 @@
-import { useOutletContext as useOutletContextRRD } from 'react-router'
-import { ethers } from 'ethers'
-
-type OutletContext = {
-  walletProvider: ethers.Eip1193Provider
-}
-
-function useOutletContext() {
-  return useOutletContextRRD<OutletContext>()
-}
-
-export { useOutletContext }
