+++ modules/passkey/README.md
@@ -25,7 +25,7 @@ Use of `SafeWebAuthnSignerProxy` provides gas savings compared to the whole cont
 
 ### [SafeWebAuthnSignerFactory](./contracts/SafeWebAuthnSignerFactory.sol)
 
-The `SafeWebAuthnSignerFactory` contract deploys the `SafeWebAuthnSignerProxy` contract with the public key coordinates and verifier information. The factory contract also supports signature verification for the public key and signature information without deploying the signer contract, which is used during the validation of ERC-4337 user operations by the experimental `SafeSignerLaunchpad` contract. Using [ISafeSignerFactory](./contracts/interfaces/ISafeSignerFactory.sol) interface and this factory contract address, new signers can be deployed.
+The `SafeWebAuthnSignerFactory` contract deploys the `SafeWebAuthnSignerProxy` contract with the public key coordinates and verifier information. The factory contract also supports signature verification for the public key and signature information without deploying the signer contract. Using [ISafeSignerFactory](./contracts/interfaces/ISafeSignerFactory.sol) interface and this factory contract address, new signers can be deployed.
 
 ### [WebAuthn](./contracts/libraries/WebAuthn.sol)
